backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.43.0
groq==1.1.2
python-multipart==0.0.22
python-dotenv==1.2.2
aiofiles==25.1.0
pydantic==2.12.5
httpx==0.28.1
